QuickContainers: refactor to enable installing as QML plugin
