+++ QuickContainers/include/qcmAdapter.h
