+++ QuickContainers/src/qcmAbstractContainer.cpp
@@ -1,0 +1,4 @@
+//
+// Created by alexander on 19/08/18.
+//
+#include "qcmAbstractContainer.h"
