+++ QuickContainers/src/qmldir
@@ -1,0 +1,3 @@
+module QuickContainers
+plugin quickcontainersplugin
+classname QuickContainersPlugin
